A restore program specify for H
A program which scans all the hard disk and find every places contain your input string. Create for H.
Using:
- Recover file by searching any part of that file which you remember. (work for files which have been damaged some part so that common recover tools could not find, but there's a hope that you can recover just some parts of it)
- Find files which you don't remember its location but its content.
- Read file/disk hexdump.
- For entertainment.
Actually, I know it's not that useful, it couldn't complete its original task which is the reason why it exists. I just finish it to escape from boredom, to be far away from reality, and to record what I thought and felt.